requirements.txt
python-shogi
PyQt6
pyinstaller
dotenv
